ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
ultralytics==8.2.50
torch==2.3.1
torchvision==0.18.1
opencv-python-headless==4.10.0.84
numpy==1.26.4
pillow==10.3.0
python-multipart==0.0.9
